refactor(plugins): Integrate MiAZImportDoc into core as a service
