+++ CHANGELOG.md
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - MiAZNotes marks documents that have notes with a dedicated leftmost workspace column (a pin icon plus the note count) instead of a CSS row highlight. The column is added to the workspace view at runtime when the plugin activates and removed when it deactivates, so it only exists while MiAZNotes is enabled. This replaces the old `.miaz-has-notes` highlight, which embedded a hardcoded absolute path to the pin SVG and so broke on any installed system; the new cell uses a theme icon name.
+- Moved `MiAZImportDoc` ("Add new document(s)") from a togglable plugin to a core service (`services/importdoc.py`, service `importdoc`, following the same pattern as `MiAZMassRename`). Every repository needs a way to add its first document, so that action should not live behind an optional, disableable plugin; a fresh repository previously had to hardcode it into its default `plugins-used.json` to avoid shipping with no way to add documents at all, which is no longer needed now that the feature cannot be disabled. The headerbar "Add" menu (`headerbar-add-menu`) now always shows the core action first, followed by any loaded Import-subcategory plugin (`MiAZAddFromDir`, `MiAZImportFromZip`, `MiAZImportFromScan`), and the Add button is now always visible instead of only when an Import plugin happened to be enabled.
 - The workspace no longer opens on an empty view when the default date filter has no documents. Right after a new month starts, "This month" is usually empty; the workspace now selects the nearest date filter that actually contains documents (walking to wider windows: "Since past month", "Since last 3 months", and so on, falling back to "All documents"). It runs only as the initial default, at load, repository switch, and the day/month rollover, and never overrides a date filter you selected by hand. The chooser (`pick_date_preset` in `widgets/workspace.py`) is covered by `tests/test_workspace_datepick.py`.
 - The rename dialog is now keyboard-friendly. It focuses the first field that needs attention when it opens, the date is an editable `YYYYMMDD` entry (the calendar remains as an assist), `Ctrl+Enter` applies from any field and `Esc` cancels, and in the concept entry `Enter` picks the highlighted autocomplete while the popover is open. Applying now refuses to build an invalid filename: when a required field (date, country, sent by, concept, sent to) is missing or invalid the dialog stays open and focuses the field to fix.
 - Merged the `MiAZAIChat` plugin into `MiAZAIAssistant`. There is now a single AI plugin that offers both features: suggesting the seven filename fields from a document (the `Suggest filename…` menu entry and the rename-dialog button) and chatting about the selected document (the `Chat with document…` menu entry). Both share one provider layer and one settings dialog, so API keys and the active provider are configured once. Internally the `miazaic` package was folded into `miazai`: every provider (Claude, OpenAI, Gemini, Ollama) now implements both `suggest()` and `chat()` on a shared base, `make_usage` has a single definition in `miazai/usage.py`, and the chat dialog moved to `miazai/ui/chat.py`. No configuration is migrated, so anyone who had configured the old `MiAZAIChat` plugin re-enters that provider's API key once. The aggregated `data/resources/plugins-requirements.json` no longer lists `miazaichat`. Covered by `tests/test_ai_plugin.py`.
