--- data/resources/plugins/MiAZImportDoc/adddoc.plugin
@@ -1,12 +1,0 @@
-[Plugin]
-Module=adddoc
-Name=MiAZImportDoc
-Loader=python
-Description=Add new document(s)
-Authors=Tomás Vírseda <tomasvirseda@gmail.com>
-Copyright=Copyright © 2025 Tomás Vírseda
-Website=http://github.com/t00m/MiAZ
-Help=http://github.com/t00m/MiAZ/README.adoc
-Version=0.1.26
-Category=Data Management
-Subcategory=Import
